Moved HasAppStrings interface from AndroidDriver to AppiumDriver to allow getting app strings in iOS
